+++ packages/apollo-cache-inmemory/src/__tests__/fragmentMatcher.ts
@@ -1,6 +1,50 @@
 import { IntrospectionFragmentMatcher } from '../fragmentMatcher';
 import { defaultNormalizedCacheFactory } from '../objectCache';
-import { ReadStoreContext } from '..';
+import { ReadStoreContext } from '../types';
+import { InMemoryCache } from '../inMemoryCache';
+import gql from 'graphql-tag';
+
+describe('FragmentMatcher', () => {
+  it('can match against the root Query', () => {
+    const cache = new InMemoryCache({
+      addTypename: true,
+    });
+
+    const query = gql`
+      query AllPeople {
+        people {
+          id
+          name
+        }
+        ...PeopleTypes
+      }
+      fragment PeopleTypes on Query {
+        __type(name: "Person") {
+          name
+          kind
+        }
+      }
+    `;
+
+    const data = {
+      people: [
+        {
+          __typename: 'Person',
+          id: 123,
+          name: 'Ben',
+        },
+      ],
+      __type: {
+        __typename: '__Type',
+        name: 'Person',
+        kind: 'OBJECT',
+      },
+    };
+
+    cache.writeQuery({ query, data });
+    expect(cache.readQuery({ query })).toEqual(data);
+  });
+});
 
 describe('IntrospectionFragmentMatcher', () => {
   it('will throw an error if match is called if it is not ready', () => {
@@ -42,12 +86,12 @@ describe('IntrospectionFragmentMatcher', () => {
       generated: false,
     };
 
-    const readStoreContext: ReadStoreContext = {
+    const readStoreContext = {
       store,
       returnPartialData: false,
       hasMissingField: false,
-      customResolvers: {},
-    } as any;
+      cacheRedirects: {},
+    } as ReadStoreContext;
 
     expect(ifm.match(idValue as any, 'Item', readStoreContext)).toBe(true);
     expect(ifm.match(idValue as any, 'NotAnItem', readStoreContext)).toBe(